:recycle: Refactor : Repository 클래스 위치 변경 - Repository 클래스 위치 변경 Related to: #181
